+++ src/components/dialog/DialogSections.tsx
@@ -1,12 +1,12 @@
 import { FC, useMemo } from "react"
 import { CSSTransition } from "react-transition-group"
+import { Line } from "../../data/line"
+import { Station } from "../../data/station"
 import img_above from "../../img/ic_above.png"
 import img_location from "../../img/map_pin.svg"
 import img_mylocation from "../../img/pin_mylocation.png"
 import img_radar from "../../img/radar.png"
 import img_station from "../../img/station.png"
-import { Line } from "../../model/line"
-import { Station } from "../../model/station"
 import { useRefCallback } from "../hooks"
 import { DialogType, StationDialogProps } from "../navState"
 import "./InfoDialog.css"
