+++ src/components/map/serviceHook.ts
@@ -1,5 +1,6 @@
 import { useDispatch } from "react-redux"
-import { Station } from "../../model/station"
+import { Station } from "../../data/station"
+import { CurrentLocation } from "../../location/location"
 import * as action from "../../redux/actions"
 import { AppDispatch } from "../../redux/store"
 import { useRefCallback } from "../hooks"
@@ -8,24 +9,19 @@ import { useRefCallback } from "../hooks"
  * serviceに登録するコールバック関数を取得する
  * @returns 
  */
-export const useServiceCallback = (progressHandler: (task: Promise<void>, text: string) => any) => {
+export const useServiceCallback = () => {
   const dispatch = useDispatch<AppDispatch>()
 
-  const onGeolocationPositionChanged = useRefCallback((pos: GeolocationPosition) => {
-    dispatch(action.setCurrentLocation(pos))
+  const onLocationChanged = useRefCallback((location: CurrentLocation) => {
+    dispatch(action.setCurrentLocation(location))
   })
 
   const onStationLoaded = useRefCallback((list: Station[]) => {
     dispatch(action.appendLoadedStation(list))
   })
 
-  const dataLoadingCallback = useRefCallback((message, promise) => {
-    progressHandler(promise, message)
-  })
-
   return {
-    onGeolocationPositionChanged,
+    onLocationChanged,
     onStationLoaded,
-    dataLoadingCallback,
   }
 }
